requirements.txt
requests
confluent-kafka
psycopg2-binary
pytz
